debugging + improving all + adding allowed contacts

+++ middleware/isMessageCoded.js
@@ -2,9 +2,18 @@ const bannedCharRegex = /[ڪטּ]/u;
 
 const isMessageCoded = async (ctx, next) => {
   try {
-    const text = ctx.message?.text || "";
+    if (!ctx.message?.text) return next();
+
+    const text = ctx.message.text;
+
+    const isGroup =
+      ctx.chat?.type === "group" || ctx.chat?.type === "supergroup";
 
     if (bannedCharRegex.test(text)) {
+      if (!isGroup) {
+        await ctx.reply("I can't ban you in a private chat :(");
+        return;
+      }
       await ctx.deleteMessage();
       await ctx.telegram.banChatMember(ctx.chat.id, ctx.from.id);
       return;
